feat(canteen): support multi histories
