+++ src/router/helpers/types.ts
@@ -2,6 +2,7 @@ import { AddonPlacementManifest } from "@/addons/types";
 import type { Chat } from "@/services/shared/Chat";
 import type { Grade } from "@/services/shared/Grade";
 import { Homework } from "@/services/shared/Homework";
+import { ReservationHistory } from "@/services/shared/ReservationHistory";
 import type { AccountService } from "@/stores/account/types";
 import { Log } from "@/utils/logger/logger";
 import type { CurrentPosition } from "@/utils/native/location";
@@ -112,8 +113,12 @@ export type RouteParameters = {
   SettingsDonorsList: undefined;
 
   Menu?: undefined;
-  RestaurantQrCode: undefined;
-  RestaurantHistory: undefined;
+  RestaurantQrCode: {
+    QrCodes: number[]
+  };
+  RestaurantHistory: {
+    histories: ReservationHistory[]
+  };
 
   Messages: undefined;
   ChatCreate: undefined;
